binance-gateway/pyproject.toml
[tool.poetry]
name = "binance-gateway"
version = "0.1.0"
description = ""
authors = ["Edwin Abot <edwinabot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-binance = "^1.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
